pyproject.toml
[project]
name = "autoimprove"
version = "0.1.0"
description = "Universal autonomous self-improvement for any software repo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "openai>=1.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
]

[project.scripts]
autoimprove = "autoimprove.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autoimprove"]

[tool.pytest.ini_options]
testpaths = ["tests"]
